+++ context_test.go
@@ -1084,7 +1084,7 @@ func TestGearContextAttachment(t *testing.T) {
 			if err != nil {
 				return err
 			}
-			return ctx.Attachment("README.md", time.Time{}, file)
+			return ctx.Attachment("Gear 设计说明.md", time.Time{}, file)
 		})
 
 		srv := app.Start()
@@ -1093,7 +1093,8 @@ func TestGearContextAttachment(t *testing.T) {
 		res, err := RequestBy("GET", "http://"+srv.Addr().String())
 		assert.Nil(err)
 		assert.Equal(200, res.StatusCode)
-		assert.Equal("attachment; filename=README.md", res.Header.Get(HeaderContentDisposition))
+		assert.Equal(`attachment; filename="Gear 设计说明.md"; filename*=UTF-8''Gear%20%E8%AE%BE%E8%AE%A1%E8%AF%B4%E6%98%8E.md`,
+			res.Header.Get(HeaderContentDisposition))
 		assert.Equal(MIMETextPlainCharsetUTF8, res.Header.Get(HeaderContentType))
 		assert.Equal(string(data), PickRes(res.Text()).(string))
 	})
@@ -1116,7 +1117,7 @@ func TestGearContextAttachment(t *testing.T) {
 		res, err := RequestBy("GET", "http://"+srv.Addr().String())
 		assert.Nil(err)
 		assert.Equal(200, res.StatusCode)
-		assert.Equal("inline; filename=README.md", res.Header.Get(HeaderContentDisposition))
+		assert.Equal(`inline; filename="README.md"`, res.Header.Get(HeaderContentDisposition))
 		assert.Equal(MIMETextPlainCharsetUTF8, res.Header.Get(HeaderContentType))
 		assert.Equal(string(data), PickRes(res.Text()).(string))
 	})