+++ CHANGELOG.md
@@ -5,4 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 -----
 
+## [1.0.1] - 2017-03-01
+
+### Fixed
+
+- Fix and improve Content-Disposition which is used by ctx.Attachment.
+- Update Readme document.
+
+-----
+
 ## [1.0.0] - 2017-03-01